refactor(patterns): split section role validation
