+++ .sampo/changesets/1029-pattern-section-role-validation.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split pattern catalog section-role marker extraction and content validation into
+a focused runtime module while keeping the public pattern catalog API stable.
